refactor: move Topbar components to its own folder
