+++ packages/website/components/Topbar/index.ts
@@ -1,0 +1,1 @@
+export { Topbar } from './Topbar';
